requirements.txt
re
praw
os
cloud-sql-python-connector[pymysql]
SQLAlchemy==1.4.39
google-cloud-storage==2.4.0